Bring over `blur` implementation.
Original implementations are from
https://github.com/cibernox/ember-native-dom-helpers and then modified:

* Added `fireEvent(element, 'focusout')`
* Remove run wrapping
